Updated tektronix for XWaveform format

+++ src/main/scala/org/odfi/instruments/nivisa/tektronix/TDS2002B.scala
@@ -54,7 +54,7 @@ class TDS2002B(baseDevice: VISAUSBDevice) extends TekTronixOsci(baseDevice) {
      //-- Save to waveform
 
      waveform.data = dataInt
-     waveform.points = points.toInt
+     waveform.points = points.toLong
      waveform.xIncrement= timeScale
      waveform.xUnit= xunit
      waveform.yIncrement=ymult
